agregando componentes y menu
